fix possible int overflow bug on meminfo for x86
change meminfo_memavailable from 'long' to 'long long'

+++ include/meminfo.h
@@ -7,7 +7,7 @@ enum    e_meminfo_param
     MEMAVAILABLE,
 };
 
-long    meminfo(enum e_meminfo_param info);
+long long   meminfo(enum e_meminfo_param info);
 
 
 #endif /* ! MEMINFO_H */
